demo/web_demo/WebRTC_Demo/omini_backend_code/code/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "minicpmo-web-backend"
version = "0.1.0"
description = "MiniCPMO Web Backend FastAPI"
authors = [
    {name = "xushihao", email = "2580507844@qq.com"}
]
readme = "README.md"
requires-python = ">=3.9.0"
dependencies = [
    # Web框架
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    # 配置
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    # LiveKit (WebRTC 核心)
    "livekit>=0.20.0",
    "livekit-agents>=0.4.0",
    "livekit-plugins-silero>=0.1.0",
    # VAD
    "webrtcvad>=2.0.10",
    # 科学计算（音视频处理必需）
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "opencv-python-headless>=4.8.0",
    "Pillow>=10.0.0",
    "soundfile>=0.12.0",
    "opuslib>=3.0.1",
    "pydub>=0.25.0",
    # 网络
    "aiohttp>=3.8.0",
    # 音频处理
    "librosa>=0.10.0",
    # 日志
    "concurrent-log-handler>=0.9.20",
]

# ===== 以下依赖已移除（本地部署不需要）=====
# "redis>=5.0.0"              -> 已用内存字典替代
# "celery>=5.3.0"             -> 未使用
# "apscheduler>=3.10.0"       -> 未使用
# "moviepy>=1.0.3"            -> 未使用
# "matplotlib>=3.7.0"         -> 未使用
# "aliyun-log-python-sdk"     -> 云服务日志，本地不需要
# "livekit-plugins-openai"    -> 未使用
# "opencv-python"             -> 改用 headless 版本（不依赖 GUI 库）

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
